extra/requirements.txt
pyyaml
redis
uuid
